+++ testomania/src/main/java/com/earth/testomania/apis/opentdb/data/source/remote/dto/OpenTdbQuizDTO.kt
@@ -1,0 +1,6 @@
+package com.earth.testomania.apis.opentdb.data.source.remote.dto
+
+data class OpenTdbQuizDTO(
+    val response_code: Int,
+    val results: List<ResultDto>
+)
