+++ testomania/src/main/java/com/earth/testomania/apis/opentdb/domain/models/OpenTdbCategory.kt
@@ -1,0 +1,27 @@
+package com.earth.testomania.apis.opentdb.domain.models
+
+enum class OpenTdbCategory(val id: Int) {
+    GENERAL(9),
+    BOOKS(10),
+    FILM(11),
+    MUSIC(12),
+    MUSICALS_AND_THEATRES(13),
+    TELEVISION(14),
+    VIDEO_GAMES(15),
+    BOARD_GAMES(16),
+    NATURE(17),
+    COMPUTERS(18),
+    MATHEMATICS(19),
+    MYTHOLOGY(20),
+    SPORTS(21),
+    GEOGRAPHY(22),
+    HISTORY(23),
+    POLITICS(24),
+    ART(25),
+    CELEBRITIES(26),
+    ANIMALS(27),
+    VEHICLES(28),
+    COMICS(29),
+    GADGETS(30),
+    ANIME_MANGA(31),
+}
