+++ testomania/src/main/java/com/earth/testomania/apis/quizapi/data/source/remote/dto/TagDTO.kt
@@ -1,0 +1,8 @@
+package com.earth.testomania.apis.quizapi.data.source.remote.dto
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class TagDTO(
+    val name: String?,
+)
