+++ testomania/src/main/java/com/earth/testomania/apis/quizapi/domain/model/AnswerKey.kt
@@ -1,0 +1,7 @@
+package com.earth.testomania.apis.quizapi.domain.model
+
+class AnswerKey {
+    companion object {
+        val list = listOf("A", "B", "C", "D", "E", "F")
+    }
+}
