+++ testomania/src/main/java/com/earth/testomania/apis/quizapi/domain/repository/QuizRepository.kt
@@ -1,0 +1,12 @@
+package com.earth.testomania.apis.quizapi.domain.repository
+
+import com.earth.testomania.apis.quizapi.domain.model.QuizCategory
+import com.earth.testomania.common.DataState
+import com.earth.testomania.common.model.Quiz
+import kotlinx.coroutines.flow.Flow
+
+interface QuizRepository {
+
+    suspend fun getQuizList(params: QuizCategory): Flow<DataState<List<Quiz>>>
+
+}
