--- testomania/src/main/java/com/earth/testomania/opentdb/models/OpenTdbQuizDTO.kt
@@ -1,6 +1,0 @@
-package com.earth.testomania.opentdb.models
-
-data class OpenTdbQuizDTO(
-    val response_code: Int,
-    val results: List<ResultDto>
-)
