--- testomania/src/main/java/com/earth/testomania/opentdb/models/ResultDto.kt
@@ -1,10 +1,0 @@
-package com.earth.testomania.opentdb.models
-
-data class ResultDto(
-    val category: String,
-    val correct_answer: String,
-    val difficulty: String,
-    val incorrect_answers: List<String>,
-    val question: String,
-    val type: String
-)
