--- testomania/src/main/java/com/earth/testomania/technical/data/source/remote/dto/TagDTO.kt
@@ -1,8 +1,0 @@
-package com.earth.testomania.technical.data.source.remote.dto
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class TagDTO(
-    val name: String?,
-)
