--- testomania/src/main/java/com/earth/testomania/technical/domain/model/AnswerKey.kt
@@ -1,7 +1,0 @@
-package com.earth.testomania.technical.domain.model
-
-class AnswerKey {
-    companion object {
-        val list = listOf("A", "B", "C", "D", "E", "F")
-    }
-}
