+++ testomania/src/main/java/com/earth/testomania/technical/presentation/ui_parts/LoadingScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
-import com.earth.testomania.technical.domain.model.TechCategory
+import com.earth.testomania.apis.quizapi.domain.model.TechCategory
 import kiwi.orbit.compose.ui.controls.LinearIndeterminateProgressIndicator
 import kotlinx.coroutines.delay
 import kotlin.time.Duration.Companion.seconds
